# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: iHR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: jobId
          in: query
          schema:
            type: ["integer", "null"]
        - name: candidateId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/send-email:
    post:
      operationId: sendApplicationEmail
      tags: [applications]
      summary: Send email to candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLog"
  /applications/{id}/links:
    patch:
      operationId: updateApplicationLinks
      tags: [applications]
      summary: Update meeting and calendar links
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationLinksUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/generate-employee:
    post:
      operationId: generateEmployeeFromApplication
      tags: [applications]
      summary: Generate employee record from a final offer application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /applications/{id}/emails:
    get:
      operationId: listApplicationEmails
      tags: [applications]
      summary: List emails sent for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailLog"
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
  /attendance/punch:
    post:
      operationId: punchAttendance
      tags: [attendance]
      summary: Punch in or out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendancePunchInput"
      responses:
        "200":
          description: Punch recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
...
components:
  schemas:
...
    Application:
      type: object
      required: [id, jobId, candidateId, status]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        status:
          type: string
          enum: [applied, ats_tracking, screening, shortlisted, interview, offer, conditional_offer, verification, final_offer, hired, rejected]
        atsScore:
          type: ["integer", "null"]
        matchPercent:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        meetingLink:
          type: ["string", "null"]
        calendarLink:
          type: ["string", "null"]
        candidateEmail:
          type: ["string", "null"]
        appliedAt:
          type: string
        interviewRound:
          type: ["integer", "null"]
        salaryOffer:
          type: ["integer", "null"]
        joiningDate:
          type: ["string", "null"]
        offerNotes:
          type: ["string", "null"]
    ApplicationInput:
      type: object
      required: [jobId, candidateId]
      properties:
        jobId:
          type: integer
        candidateId:
          type: integer
        coverLetter:
          type: ["string", "null"]
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        interviewRound:
          type: ["integer", "null"]
        salaryOffer:
          type: ["integer", "null"]
        joiningDate:
          type: ["string", "null"]
        offerNotes:
          type: ["string", "null"]
...
    SendEmailInput:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
        body:
          type: string
    ApplicationLinksUpdate:
      type: object
      properties:
        meetingLink:
          type: ["string", "null"]
        calendarLink:
          type: ["string", "null"]
    EmailLog:
      type: object
      required: [id, applicationId, toEmail, subject, body, sentAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        toEmail:
          type: string
        subject:
          type: string
        body:
          type: string
        sentBy:
          type: ["string", "null"]
        sentAt:
          type: string
...
    Employee:
      type: object
      required: [id, employeeCode, fullName, companyId, department, status]
      properties:
        id:
          type: integer
        employeeCode:
          type: string
        fullName:
          type: string
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: string
        companyId:
          type: integer
        department:
          type: string
        designation:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        reportingManager:
          type: ["string", "null"]
        reportingManagerId:
          type: ["integer", "null"]
        salary:
          type: ["integer", "null"]
        basicSalary:
          type: ["integer", "null"]
        salaryUpdatedDate:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        positionUpdatedDate:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        doe:
          type: ["string", "null"]
        ofcEmail:
          type: ["string", "null"]
        ofcNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave]
        avatarUrl:
          type: ["string", "null"]
        leaveBalance:
          type: ["integer", "null"]
...
    Attendance:
      type: object
      required: [id, employeeId, date, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        date:
          type: string
        punchIn:
          type: ["string", "null"]
        punchOut:
          type: ["string", "null"]
        status:
          type: string
          enum: [present, absent, half_day, late]
        hoursWorked:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
    AttendancePunchInput:
      type: object
      required: [employeeId, type]
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [punch_in, punch_out]
        location:
          type: ["string", "null"]
...